refactor: fd_check.py → tangent.py にリネーム
jacobian.py との命名規則を統一。モジュール名がクラス (TangentChecker) の
実装手段 (fd) ではなくコンセプト (tangent verification) を示すように変更。
