+++ tests/unit/test_stress_field.py
@@ -20,7 +20,7 @@ from manforge.core.dimension import SOLID_3D
 from manforge.models.j2_isotropic import J2Isotropic3D
 from manforge.models.ow_kinematic import OWKinematic3D
 from manforge.simulation.integrator import PythonIntegrator
-from manforge.verification.fd_check import check_tangent
+from manforge.verification.tangent import check_tangent
 
 
 _J2_PARAMS = dict(E=210000.0, nu=0.3, sigma_y0=250.0, H=1000.0)
